# ipx/docs/swagger.yaml
basePath: /
definitions:
...
  v3.BuildERC20EIP1559TransactionRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      contract:
        example: 0x5FbDB2315678afecb367f032d93F642f64180aa3
        type: string
      from:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
      to:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  v3.BuildERC20EIP1559TransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
  v3.BuildERC20LegacyTransactionRequest:
    properties:
      amount:
        example: "1000000000000000000"
        type: string
      contract:
        example: 0x5FbDB2315678afecb367f032d93F642f64180aa3
        type: string
      from:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
      to:
        example: 0xDa70aA79f1a329719b9cf9d334b0a82b1d5269f3
        type: string
    type: object
  v3.BuildERC20LegacyTransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
  v3.BuildNativeEIP1559TransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        description: wei as decimal string
        type: string
    type: object
  v3.BuildNativeEIP1559TransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
  v3.BuildNativeLegacyTransactionRequest:
    properties:
      from:
        type: string
      to:
        type: string
      value:
        description: wei as decimal string
        type: string
    type: object
  v3.BuildNativeLegacyTransactionResponse:
    properties:
      r:
        type: string
      s:
        type: string
      signed_rlp:
        type: string
      tx_hash:
        type: string
      unsigned_rlp:
        type: string
      v:
        type: string
    type: object
...
host: localhost:33152
info:
  contact: {}
  description: Ethereum transaction API
  title: evmlab API
  version: "1.0"
paths:
...
  /evm/v3/transaction/erc20/eip1559:
    post:
      consumes:
      - application/json
      description: Builds transfer(address,uint256) calldata, estimates gas, signs
        with the configured key, returns unsigned_rlp, signed_rlp, and tx_hash
      parameters:
      - description: ERC-20 transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v3.BuildERC20EIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.BuildERC20EIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build and sign ERC-20 EIP-1559 transfer tx
      tags:
      - transaction
  /evm/v3/transaction/erc20/legacy:
    post:
      consumes:
      - application/json
      description: Builds transfer(address,uint256) calldata, estimates gas, signs
        with the configured key, returns unsigned_rlp, signed_rlp, and tx_hash
      parameters:
      - description: ERC-20 transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v3.BuildERC20LegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.BuildERC20LegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build and sign ERC-20 legacy transfer tx
      tags:
      - transaction
  /evm/v3/transaction/native/eip1559:
    post:
      consumes:
      - application/json
      description: Fetches chain state, builds and signs an EIP-1559 native transfer,
        returns unsigned_rlp, signed_rlp, and tx_hash
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v3.BuildNativeEIP1559TransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.BuildNativeEIP1559TransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build and sign EIP-1559 transfer tx
      tags:
      - transaction
  /evm/v3/transaction/native/legacy:
    post:
      consumes:
      - application/json
      description: Fetches chain state, builds and signs an EIP-155 legacy native
        transfer, returns unsigned_rlp, signed_rlp, and tx_hash
      parameters:
      - description: Transfer request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v3.BuildNativeLegacyTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v3.BuildNativeLegacyTransactionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Build and sign legacy transfer tx
      tags:
      - transaction
...
swagger: "2.0"